Clean up the output file
This is a userscript, which have essentially unrestricted access to the
user's browser. Therefore, it should be easy for the user to review the
code they are about to install.

This commit adds several Rollup plugins designed to make the output file
easier to read, including adding comments to the output to show what the
original input files were and reformatting the code using Prettier.